requirements.txt
alembic==1.6.5
altgraph==0.17
aniso8601==9.0.1
astroid==2.4.2
autopep8==1.5.4
backports.entry-points-selectable==1.1.0
baidu-aip==2.2.18.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
distlib==0.3.2
docopt==0.6.2
dominate==2.6.0
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.15.1
future==0.18.2
gevent==20.12.1
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
isort==5.6.4
itchat==1.3.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonpath==0.82
lazy-object-proxy==1.4.3
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
openpyxl==3.0.5
pbr==5.6.0
pefile==2019.4.18
Pillow==8.4.0
pipreqs==0.4.10
platformdirs==2.0.2
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
pylint==2.6.0
pymongo==3.11.2
PyMySQL==0.10.1
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.22
stevedore==3.3.0
toml==0.10.2
urllib3==1.26.2
virtualenv==20.6.0
virtualenv-clone==0.5.5
virtualenvwrapper-win==1.2.6
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.2.0
